# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "description": "ShareMore is Social Media Platform API designed to facilitate seamless user interaction through posts, media sharing, likes, and comments.",
    "title": "ShareMore"
  },
  "servers": [
    {
      "url": "http://localhost:3100"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    }
  },
  "paths": {
...
    "/api/posts": {
      "post": {
        "tags": ["Posts"],
        "summary": "Create a Post",
        "security": [
          {
            "JWT": []
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "caption": {
                    "type": "string"
                  },
                  "imageUrl": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          },
          "description": "Enter Caption and file to create a post"
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad request"
          }
        }
      },
      "get": {
        "tags": ["Posts"],
        "summary": "Get user posts",
        "description": "Get post based on user credentials",
        "security": [
          {
            "JWT": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}